pyproject.toml
[project]
name = "encumbered-twitter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.0.1",
    "selenium>=4.25.0",
]
